Moved map files into separate package.
