+++ app/src/main/res/layout/mapviewer.xml
@@ -5,7 +5,7 @@
         android:layout_height="match_parent"
         >
 
-        <de.storchp.opentracks.osmplugin.MapsforgeMapView
+        <de.storchp.opentracks.osmplugin.maps.MapsforgeMapView
             android:id="@+id/mapView"
             android:layout_width="fill_parent"
             android:layout_height="wrap_content"
